requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-huggingface==0.0.3

faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0

sentence-transformers==5.2.2
transformers>=4.45.0
huggingface_hub>=0.26.0

groq==0.37.1
streamlit==1.37.1
